backend/requirements.txt
langchain_core
python-dotenv
langchain_openai
termcolor
fastapi
uvicorn